Move logic to custom hook
